Separate View interfaces and move to nunit.gui.core
