+++ src/nunit-gui/Views/AddinPages/AddinsView.cs
