--- src/nunit-gui/Views/IAddinsView.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using NUnit.Engine.Extensibility;
-
-namespace NUnit.Gui.Views
-{
-    public interface IAddinsView : IDialog
-    {
-        void AddExtensionPoint(IExtensionPoint extensionPoint);
-    }
-}
