+++ src/nunit-gui/Views/StatusBarView.cs
@@ -25,20 +25,6 @@ using System.Windows.Forms;
 
 namespace NUnit.Gui.Views
 {
-    // Interface is used by presenter and tests
-    public interface IStatusBarView : IView
-    {
-        void OnTestLoaded(int testCount);
-        void OnTestUnloaded();
-        void OnRunStarting(int testCount);
-        void OnRunFinished(double elapsedTime);
-        void OnTestStarting(string name);
-        void OnTestPassed();
-        void OnTestFailed();
-        void OnTestWarning();
-        void OnTestInconclusive();
-    }
-
     public partial class StatusBarView : UserControl, IStatusBarView
     {
         // Counters are maintained in the view even though
