+++ src/nunit-gui/Views/XmlView.cs
@@ -28,21 +28,6 @@ using NUnit.UiKit.Elements;
 
 namespace NUnit.Gui.Views
 {
-    public interface IXmlView : IView
-    {
-        bool Visible { get; set; }
-        string Header { get; set; }
-        IViewElement XmlPanel { get; }
-        IControlElement<RichTextBox> XmlTextBox { get; }
-        IMenu CopyToolStripMenuItem { get; }
-        IMenu WordWrapToolStripMenuItem { get; }
-        XmlNode TestXml { get; set; }
-        event CommandHandler SelectAllCommand;
-        event CommandHandler SelectionChanged;
-        event CommandHandler CopyCommand;
-        event CommandHandler WordWrapChanged;
-    }
-
     public partial class XmlView : UserControl, IXmlView
     {
         private XmlNode _testXml;
