+++ src/nunit.gui.core/Views/IDialog.cs
