+++ src/nunit.gui.core/Views/IDialogManager.cs
@@ -21,8 +21,6 @@
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 // ***********************************************************************
 
-using System;
-
 namespace NUnit.Gui.Views
 {
     public interface IDialogManager
