+++ src/nunit.gui.core/Views/IMainView.cs
